NUCLEO_F411RE: Use default implementation of NV Seed
Pretend to be a PSA target so that the default NV Seed implementation
provided for use by PSA targets will be available for use on the
NUCLEO_F411RE, a non-PSA target that also lacks a TRNG.

Note that enabling this requires the device to be manufactured with an
external entropy source to be used as the initial NV Seed. Any use of
cryptography will fail at run-time until the initial NV Seed is
injected.